Day4 solution
Part1 much harder than part 2.  Not sure if there is a list comprehension for contains all.  There is one for contains any, which is in part 2.  Used a subtraction, but had to compare both list1 in list2 and list2 in list1, so needed two lists for each as the contains does a remove to compare.